+++ solr/__init__.py
@@ -5,6 +5,7 @@ from .exceptions import SolrException, SolrVersionError
 from .response import Response, Results
 from .parsers import parse_json_response, parse_query_response
 from .utils import read_response, check_response_status
+from .transport import SolrTransport
 from .schema import SchemaAPI
 from .mlt import MoreLikeThis
 from .suggest import Suggest